tint/writer/glsl: Inline constant expressions
This is required to handle materialized values, and for constant
expressions.

Bug: tint:1504
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/92080
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>

+++ test/tint/bug/tint/534.wgsl.expected.glsl
@@ -48,9 +48,9 @@ void tint_symbol_1(uvec3 GlobalInvocationID) {
   }
   uint outputIndex = ((GlobalInvocationID.y * uint(size.x)) + GlobalInvocationID.x);
   if (success) {
-    tint_symbol.result[outputIndex] = uint(1);
+    tint_symbol.result[outputIndex] = 1u;
   } else {
-    tint_symbol.result[outputIndex] = uint(0);
+    tint_symbol.result[outputIndex] = 0u;
   }
 }
